refactor(types): move and export types from package
