+++ src/index.ts
@@ -1,5 +1,39 @@
 import type { z } from "zod";
-import type { SafeActionOverload } from "./types";
+
+/*********************** TYPES ***********************/
+
+// The type for client action, which is called by components.
+// You pass the input data here, and it's all typesafe.
+export type ClientAction<IV extends z.ZodTypeAny, AO> = (input: z.input<IV>) => Promise<{
+	data?: AO;
+	serverError?: true;
+	validationError?: Partial<Record<keyof z.input<IV>, string[]>>;
+}>;
+
+// We need to overload the `safeAction` function, because some actions
+// need authentication, and others don't, so you can pass the `withAuth: true` property
+// in the `opts` arg, to get back both `parsedInput` and `authArgs` in the server
+// action function definition.
+// `authArgs` comes from the previously defined `getAuthData` function.
+export type SafeActionOverload<AuthData extends object> = {
+	<const IV extends z.ZodTypeAny, const AO>(
+		opts: {
+			input: IV;
+			withAuth?: false;
+		},
+		actionDefinition: (parsedInput: z.input<IV>, authArgs: undefined) => Promise<AO>
+	): ClientAction<IV, AO>;
+
+	<const IV extends z.ZodTypeAny, const AO>(
+		opts: {
+			input: IV;
+			withAuth: true;
+		},
+		actionDefinition: (parsedInput: z.input<IV>, authArgs: AuthData) => Promise<AO>
+	): ClientAction<IV, AO>;
+};
+
+/*********************** FUNCTIONS ***********************/
 
 // This is the safe action initializer.
 export const createSafeActionClient = <AuthData extends object>(createOpts?: {
